pyproject.toml
[tool.poetry]
authors = ["Jack Cho <jack06215@gmail.com>"]
description = ""
name = "pyton-monorepo"
packages = [
  {include = "shared_module"},
]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
Markdown = "^3.7"
PyJWT = "^2.8.0"
adlfs = "^2024.4.1"
asn1crypto = "^1.5.1"
auth0-python = "^4.7.1"
beautifulsoup4 = "^4.12.3"
cffi = "^1.16.0"
colorlog = "^6.8.2"
cryptography = "^42.0.8"
dependency-injector = "^4.41.0"
duckdb = "^1.0.0"
duckduckgo-search = "^6.2.6"
elasticsearch = "^8.14.0"
elasticsearch-async = "^6.2.0"
email-validator = "^2.2.0"
faiss-cpu = "^1.8.0.post1"
fake-useragent = "^1.5.1"
fastapi = "^0.111.0"
fastparquet = "^2024.5.0"
google-api-python-client = "^2.142.0"
google-auth = "^2.34.0"
google-auth-oauthlib = "^1.2.1"
gspread = "^6.1.2"
gspread-formatting = "^1.2.0"
idna = "^3.7"
instructor = "^1.3.4"
langchain = "^0.2.7"
langchain-community = "^0.2.7"
langchain-experimental = "^0.0.62"
langchain-openai = "^0.1.16"
langgraph = "^0.1.8"
mesop = "^0.12.2"
notebook = "^7.2.1"
numpy = "1.26.4"
openai = "^1.40.1"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"
pyarrow = "^16.0.0"
pycparser = "^2.22"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
python = "3.11.7"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
six = "^1.16.0"
slack-sdk = "^3.31.0"
sqlalchemy = "^2.0.31"
sseclient-py = "^1.8.0"
tenacity = "^8.5.0"
termcolor = "^2.4.0"
tqdm = "^4.66.4"
typeguard = "^4.3.0"
xlsxwriter = "^3.2.0"

[tool.poetry.dev-dependencies]
autoflake = "^2.3.1"
autopep8 = "^2.3.1"
flake8 = "^7.1.0"
isort = "^5.13.2"
pandas-stubs = "^2.2.2.240603"
pre-commit = "^3.7.1"
types-requests = "^2.32.0.20240712"
google-api-python-client-stubs = "^1.27.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
